hackaton/requirements.txt
django==3.0.3
django-qr-code==1.1.0
# tensorflow
# Pillow
googletrans
